fix : useSearchParam Suspense로 안 감싸줘서 발생한 에러 해결
Content 컴포넌트 생성해서 해당 부분에 모든 컴포넌트 넣음.
